Make verification less prominent
Add the hover to the listing instead of showing the link, follows the
proposal we worked on at the mobile hackfest

Closes #1270
